# Store-native TTL management