--- a/pstoreds/addr_book.go
+++ b/pstoreds/addr_book.go
@@ -1,24 +1,33 @@
 package pstoreds
 
 import (
+	"bytes"
 	"context"
+	"encoding/binary"
+	"errors"
 	"sync"
 	"time"
 
-	lru "github.com/hashicorp/golang-lru"
+	"github.com/hashicorp/golang-lru"
 	ds "github.com/ipfs/go-datastore"
-	query "github.com/ipfs/go-datastore/query"
+	"github.com/ipfs/go-datastore/query"
 	logging "github.com/ipfs/go-log"
-	peer "github.com/libp2p/go-libp2p-peer"
+	"github.com/libp2p/go-libp2p-peer"
 	ma "github.com/multiformats/go-multiaddr"
 	mh "github.com/multiformats/go-multihash"
 
 	pstore "github.com/libp2p/go-libp2p-peerstore"
-	pstoremem "github.com/libp2p/go-libp2p-peerstore/pstoremem"
+	"github.com/libp2p/go-libp2p-peerstore/pstoremem"
 )
 
 var (
 	log = logging.Logger("peerstore/ds")
+	// The maximum representable value in time.Time is time.Unix(1<<63-62135596801, 999999999).
+	// But it's too brittle and implementation-dependent, so we prefer to use 1<<62, which is in the
+	// year 146138514283. We're safe.
+	maxTime = time.Unix(1<<62, 0)
+
+	ErrTTLDatastore = errors.New("datastore must provide TTL support")
 )
 
 var _ pstore.AddrBook = (*dsAddrBook)(nil)
@@ -28,15 +37,30 @@ var _ pstore.AddrBook = (*dsAddrBook)(nil)
 type dsAddrBook struct {
 	cache        cache
 	ds           ds.TxnDatastore
-	ttlManager   *ttlManager
 	subsManager  *pstoremem.AddrSubManager
 	writeRetries int
+}
+
+type ttlWriteMode int
+
+const (
+	ttlOverride ttlWriteMode = iota
+	ttlExtend
+)
+
+type cacheEntry struct {
+	expiration time.Time
+	addrs      []ma.Multiaddr
 }
 
 // NewAddrBook initializes a new address book given a
 // Datastore instance, a context for managing the TTL manager,
 // and the interval at which the TTL manager should sweep the Datastore.
-func NewAddrBook(ctx context.Context, ds ds.TxnDatastore, opts PeerstoreOpts) (*dsAddrBook, error) {
+func NewAddrBook(ctx context.Context, store ds.TxnDatastore, opts PeerstoreOpts) (*dsAddrBook, error) {
+	if _, ok := store.(ds.TTLDatastore); !ok {
+		return nil, ErrTTLDatastore
+	}
+
 	var (
 		cache cache = &noopCache{}
 		err   error
@@ -50,8 +74,7 @@ func NewAddrBook(ctx context.Context, ds ds.TxnDatastore, opts PeerstoreOpts) (*
 
 	mgr := &dsAddrBook{
 		cache:        cache,
-		ds:           ds,
-		ttlManager:   newTTLManager(ctx, ds, &cache, opts.TTLInterval),
+		ds:           store,
 		subsManager:  pstoremem.NewAddrSubManager(),
 		writeRetries: int(opts.WriteRetries),
 	}
@@ -60,7 +83,7 @@ func NewAddrBook(ctx context.Context, ds ds.TxnDatastore, opts PeerstoreOpts) (*
 
 // Stop will signal the TTL manager to stop and block until it returns.
 func (mgr *dsAddrBook) Stop() {
-	mgr.ttlManager.cancel()
+	// noop
 }
 
 func keysAndAddrs(p peer.ID, addrs []ma.Multiaddr) ([]ds.Key, []ma.Multiaddr, error) {
@@ -98,7 +121,7 @@ func (mgr *dsAddrBook) AddAddrs(p peer.ID, addrs []ma.Multiaddr, ttl time.Durati
 	if ttl <= 0 {
 		return
 	}
-	mgr.setAddrs(p, addrs, ttl, false)
+	mgr.setAddrs(p, addrs, ttl, ttlExtend)
 }
 
 // SetAddr will add or update the TTL of an address in the AddrBook.
@@ -113,7 +136,7 @@ func (mgr *dsAddrBook) SetAddrs(p peer.ID, addrs []ma.Multiaddr, ttl time.Durati
 		mgr.deleteAddrs(p, addrs)
 		return
 	}
-	mgr.setAddrs(p, addrs, ttl, true)
+	mgr.setAddrs(p, addrs, ttl, ttlOverride)
 }
 
 func (mgr *dsAddrBook) deleteAddrs(p peer.ID, addrs []ma.Multiaddr) error {
@@ -137,11 +160,10 @@ func (mgr *dsAddrBook) deleteAddrs(p peer.ID, addrs []ma.Multiaddr) error {
 		return err
 	}
 
-	mgr.ttlManager.deleteTTLs(keys)
 	return nil
 }
 
-func (mgr *dsAddrBook) setAddrs(p peer.ID, addrs []ma.Multiaddr, ttl time.Duration, ttlReset bool) error {
+func (mgr *dsAddrBook) setAddrs(p peer.ID, addrs []ma.Multiaddr, ttl time.Duration, mode ttlWriteMode) error {
 	// Keys and cleaned up addresses.
 	keys, addrs, err := keysAndAddrs(p, addrs)
 	if err != nil {
@@ -149,10 +171,11 @@ func (mgr *dsAddrBook) setAddrs(p peer.ID, addrs []ma.Multiaddr, ttl time.Durati
 	}
 
 	mgr.cache.Remove(p.Pretty())
+
 	// Attempt transactional KV insertion.
 	var existed []bool
 	for i := 0; i < mgr.writeRetries; i++ {
-		if existed, err = mgr.dbInsert(keys, addrs); err == nil {
+		if existed, err = mgr.dbInsert(keys, addrs, ttl, mode); err == nil {
 			break
 		}
 		log.Errorf("failed to write addresses for peer %s: %s\n", p.Pretty(), err)
@@ -170,41 +193,54 @@ func (mgr *dsAddrBook) setAddrs(p peer.ID, addrs []ma.Multiaddr, ttl time.Durati
 		}
 	}
 
-	// Force update TTLs only if TTL reset was requested; otherwise
-	// insert the appropriate TTL entries if they don't already exist.
-	if ttlReset {
-		mgr.ttlManager.setTTLs(keys, ttl)
-	} else {
-		mgr.ttlManager.insertTTLs(keys, ttl)
-	}
-
 	return nil
 }
 
 // dbInsert performs a transactional insert of the provided keys and values.
-func (mgr *dsAddrBook) dbInsert(keys []ds.Key, addrs []ma.Multiaddr) ([]bool, error) {
+func (mgr *dsAddrBook) dbInsert(keys []ds.Key, addrs []ma.Multiaddr, ttl time.Duration, mode ttlWriteMode) ([]bool, error) {
 	var (
 		err     error
 		existed = make([]bool, len(keys))
+		exp     = time.Now().Add(ttl)
+		ttlB    = make([]byte, 8)
 	)
+
+	binary.LittleEndian.PutUint64(ttlB, uint64(ttl))
 
 	txn := mgr.ds.NewTransaction(false)
 	defer txn.Discard()
 
+	ttltxn := txn.(ds.TTLDatastore)
 	for i, key := range keys {
 		// Check if the key existed previously.
-		if existed[i], err = txn.Has(key); err != nil {
+		if existed[i], err = ttltxn.Has(key); err != nil {
 			log.Errorf("transaction failed and aborted while checking key existence: %s, cause: %v", key.String(), err)
 			return nil, err
 		}
 
-		// The key embeds a hash of the value, so if it existed, we can safely skip the insert.
+		// The key embeds a hash of the value, so if it existed, we can safely skip the insert and
+		// just update the TTL.
 		if existed[i] {
+			switch mode {
+			case ttlOverride:
+				err = ttltxn.SetTTL(key, ttl)
+			case ttlExtend:
+				var curr time.Time
+				if curr, err = ttltxn.GetExpiration(key); err != nil && exp.After(curr) {
+					err = ttltxn.SetTTL(key, ttl)
+				}
+			}
+			if err != nil {
+				// mode will be printed as an int
+				log.Errorf("failed while updating the ttl for key: %s, mode: %v, cause: %v", key.String(), mode, err)
+				return nil, err
+			}
 			continue
 		}
 
-		// Attempt to add the key.
-		if err = txn.Put(key, addrs[i].Bytes()); err != nil {
+		// format: bytes(ttl) || bytes(multiaddr)
+		value := append(ttlB, addrs[i].Bytes()...)
+		if err = ttltxn.PutWithTTL(key, value, ttl); err != nil {
 			log.Errorf("transaction failed and aborted while setting key: %s, cause: %v", key.String(), err)
 			return nil, err
 		}
@@ -221,25 +257,81 @@ func (mgr *dsAddrBook) dbInsert(keys []ds.Key, addrs []ma.Multiaddr) ([]bool, er
 // UpdateAddrs will update any addresses for a given peer and TTL combination to
 // have a new TTL.
 func (mgr *dsAddrBook) UpdateAddrs(p peer.ID, oldTTL time.Duration, newTTL time.Duration) {
-	prefix := ds.NewKey(p.Pretty())
-	mgr.ttlManager.adjustTTLs(prefix, oldTTL, newTTL)
+	mgr.cache.Remove(p.Pretty())
+
+	var err error
+	for i := 0; i < mgr.writeRetries; i++ {
+		if err = mgr.dbUpdateTTL(p, oldTTL, newTTL); err == nil {
+			break
+		}
+		log.Errorf("failed to update ttlsfor peer %s: %s\n", p.Pretty(), err)
+	}
+
+	if err != nil {
+		log.Errorf("failed to avoid write conflict when updating ttls for peer %s after %d retries: %v\n",
+			p.Pretty(), mgr.writeRetries, err)
+	}
+}
+
+func (mgr *dsAddrBook) dbUpdateTTL(p peer.ID, oldTTL time.Duration, newTTL time.Duration) error {
+	var (
+		prefix     = ds.NewKey(p.Pretty())
+		q          = query.Query{Prefix: prefix.String(), KeysOnly: false}
+		oldb, newb = make([]byte, 8), make([]byte, 8)
+		results    query.Results
+		err        error
+	)
+
+	binary.LittleEndian.PutUint64(oldb, uint64(oldTTL))
+	binary.LittleEndian.PutUint64(newb, uint64(newTTL))
+
+	txn := mgr.ds.NewTransaction(false)
+	defer txn.Discard()
+
+	if results, err = txn.Query(q); err != nil {
+		return err
+	}
+	defer results.Close()
+
+	ttltxn := txn.(ds.TTLDatastore)
+	for result := range results.Next() {
+		// format: bytes(ttl) || bytes(multiaddr)
+		if curr := result.Value[:8]; !bytes.Equal(curr, oldb) {
+			continue
+		}
+		newVal := append(newb, result.Value[8:]...)
+		if err = ttltxn.PutWithTTL(ds.RawKey(result.Key), newVal, newTTL); err != nil {
+			return err
+		}
+	}
+
+	if err := txn.Commit(); err != nil {
+		log.Errorf("failed to commit transaction when updating ttls, cause: %v", err)
+		return err
+	}
+
+	return nil
 }
 
 // Addrs returns all of the non-expired addresses for a given peer.
 func (mgr *dsAddrBook) Addrs(p peer.ID) []ma.Multiaddr {
 	var (
 		prefix  = ds.NewKey(p.Pretty())
-		q       = query.Query{Prefix: prefix.String(), KeysOnly: false}
+		q       = query.Query{Prefix: prefix.String(), KeysOnly: false, ReturnExpirations: true}
 		results query.Results
 		err     error
 	)
 
-	// Check the cache.
-	if entry, ok := mgr.cache.Get(p.Pretty()); ok {
-		e := entry.([]ma.Multiaddr)
-		addrs := make([]ma.Multiaddr, len(e))
-		copy(addrs, e)
-		return addrs
+	// Check the cache and return the entry only if it hasn't expired; if expired, remove.
+	if e, ok := mgr.cache.Get(p.Pretty()); ok {
+		entry := e.(cacheEntry)
+		if entry.expiration.After(time.Now()) {
+			addrs := make([]ma.Multiaddr, len(entry.addrs))
+			copy(addrs, entry.addrs)
+			return addrs
+		} else {
+			mgr.cache.Remove(p.Pretty())
+		}
 	}
 
 	txn := mgr.ds.NewTransaction(true)
@@ -252,16 +344,24 @@ func (mgr *dsAddrBook) Addrs(p peer.ID) []ma.Multiaddr {
 	defer results.Close()
 
 	var addrs []ma.Multiaddr
+	// used to set the expiration for the entire cache entry
+	earliestExp := maxTime
 	for result := range results.Next() {
-		if addr, err := ma.NewMultiaddrBytes(result.Value); err == nil {
+		// extract multiaddr from value: bytes(ttl) || bytes(multiaddr)
+		if addr, err := ma.NewMultiaddrBytes(result.Value[8:]); err == nil {
 			addrs = append(addrs, addr)
+		}
+
+		if exp := result.Expiration; !exp.IsZero() && exp.Before(earliestExp) {
+			earliestExp = exp
 		}
 	}
 
 	// Store a copy in the cache.
 	addrsCpy := make([]ma.Multiaddr, len(addrs))
 	copy(addrsCpy, addrs)
-	mgr.cache.Add(p.Pretty(), addrsCpy)
+	entry := cacheEntry{addrs: addrsCpy, expiration: earliestExp}
+	mgr.cache.Add(p.Pretty(), entry)
 
 	return addrs
 }
@@ -320,7 +420,7 @@ func (mgr *dsAddrBook) ClearAddrs(p peer.ID) {
 
 	if e, ok := mgr.cache.Peek(p.Pretty()); ok {
 		mgr.cache.Remove(p.Pretty())
-		keys, _, _ := keysAndAddrs(p, e.([]ma.Multiaddr))
+		keys, _, _ := keysAndAddrs(p, e.(cacheEntry).addrs)
 		deleteFn = func() error {
 			return mgr.dbDelete(keys)
 		}
@@ -342,9 +442,6 @@ func (mgr *dsAddrBook) ClearAddrs(p peer.ID) {
 	if err != nil {
 		log.Errorf("failed to clear addresses for peer %s after %d attempts\n", p.Pretty(), mgr.writeRetries)
 	}
-
-	// Perform housekeeping.
-	mgr.ttlManager.clear(prefix)
 }
 
 // dbDelete transactionally deletes the provided keys.
@@ -495,7 +592,7 @@ func (mgr *ttlManager) deleteTTLs(keys []ds.Key) {
 	}
 }
 
-func (mgr *ttlManager) insertTTLs(keys []ds.Key, ttl time.Duration) {
+func (mgr *ttlManager) bumpTTLs(keys []ds.Key, ttl time.Duration) {
 	mgr.Lock()
 	defer mgr.Unlock()