**This PR replaces the bespoke TTL manager with store-native TTL management.**

Notes:

* Entries are now expired by the DB without notifying us. To keep the cache consistent, we now store the invalidation timestamp, calculated as the _earliest_ expiration out of all multiaddrs. Before we return a cached entry, we check it's expiration and remove it if it's expired, falling through to the DB to repopulate the entry afresh.

* To support the `UpdateAddrs()` operation, which performs a conditional TTL update, we need to store the original TTL alongside the entry. Since the requirement is simple, I have not used anything fancy for sedes. I'm just packing the TTL (int64) + entry, by concatenating their byte strings.

* Cache keys are now `peer.ID`, no longer stringing them.

* Our friend badger doesn't support sub-second TTLs, so I've had to adjust the time-dependent TTL tests to use above-second resolutions.

* Last but not least, previously we had two methods on the `ttlManager` to update the expiration: `insertOrExtendTTL` and `setTTL`:
  * The former inserted TTLs for new keys, or updated existing ones _only to extend their TTL_, it would never bring the expiration closer. This is now called `ttlExtend` mode.
  * The latter updated the TTL always. This is now called `ttlOverride` mode.

Both modes are represented by enum values.