@raulk 2018-09-14
Benchmark comparison between master and this branch.

```
→ benchcmp txn.txt ttl.txt
benchmark                                                                 old ns/op     new ns/op     delta
BenchmarkBadgerDsPeerstore/AddAddrs-1Addrs-Caching-8                      125522        119486        -4.81%
BenchmarkBadgerDsPeerstore/AddAddrs-10Addrs-Caching-8                     262138        250963        -4.26%
BenchmarkBadgerDsPeerstore/AddAddrs-100Addrs-Caching-8                    760962        750596        -1.36%
BenchmarkBadgerDsPeerstore/SetAddrs-1Addrs-Caching-8                      116322        104276        -10.36%
BenchmarkBadgerDsPeerstore/SetAddrs-10Addrs-Caching-8                     251848        246731        -2.03%
BenchmarkBadgerDsPeerstore/SetAddrs-100Addrs-Caching-8                    797630        742194        -6.95%
BenchmarkBadgerDsPeerstore/GetAddrs-1Addrs-Caching-8                      2848          1694          -40.52%
BenchmarkBadgerDsPeerstore/GetAddrs-10Addrs-Caching-8                     2912          1765          -39.39%
BenchmarkBadgerDsPeerstore/GetAddrs-100Addrs-Caching-8                    3695          2140          -42.08%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-1Addrs-Caching-8           806264        1119209       +38.81%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-10Addrs-Caching-8          3610875       3405595       -5.69%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-100Addrs-Caching-8         9733449       9621773       -1.15%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-1Addrs-Caching-8         5729764       6086530       +6.23%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-10Addrs-Caching-8        15502463      18584244      +19.88%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-100Addrs-Caching-8       135901828     160347525     +17.99%
BenchmarkBadgerDsPeerstore/AddAddrs-1Addrs-Cacheless-8                    117767        105665        -10.28%
BenchmarkBadgerDsPeerstore/AddAddrs-10Addrs-Cacheless-8                   253219        246950        -2.48%
BenchmarkBadgerDsPeerstore/AddAddrs-100Addrs-Cacheless-8                  760662        748213        -1.64%
BenchmarkBadgerDsPeerstore/SetAddrs-1Addrs-Cacheless-8                    121158        104135        -14.05%
BenchmarkBadgerDsPeerstore/SetAddrs-10Addrs-Cacheless-8                   293593        246731        -15.96%
BenchmarkBadgerDsPeerstore/SetAddrs-100Addrs-Cacheless-8                  758146        729168        -3.82%
BenchmarkBadgerDsPeerstore/GetAddrs-1Addrs-Cacheless-8                    30867         22391         -27.46%
BenchmarkBadgerDsPeerstore/GetAddrs-10Addrs-Cacheless-8                   82003         71262         -13.10%
BenchmarkBadgerDsPeerstore/GetAddrs-100Addrs-Cacheless-8                  253858        514980        +102.86%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-1Addrs-Cacheless-8         1289891       1244353       -3.53%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-10Addrs-Cacheless-8        6629343       5758940       -13.13%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-100Addrs-Cacheless-8       13702590      12800675      -6.58%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-1Addrs-Cacheless-8       5578808       6106094       +9.45%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-10Addrs-Cacheless-8      15358239      18367850      +19.60%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-100Addrs-Cacheless-8     130995339     167926230     +28.19%

benchmark                                                                 old allocs     new allocs     delta
BenchmarkBadgerDsPeerstore/AddAddrs-1Addrs-Caching-8                      138            124            -10.14%
BenchmarkBadgerDsPeerstore/AddAddrs-10Addrs-Caching-8                     677            673            -0.59%
BenchmarkBadgerDsPeerstore/AddAddrs-100Addrs-Caching-8                    6099           6092           -0.11%
BenchmarkBadgerDsPeerstore/SetAddrs-1Addrs-Caching-8                      127            124            -2.36%
BenchmarkBadgerDsPeerstore/SetAddrs-10Addrs-Caching-8                     677            673            -0.59%
BenchmarkBadgerDsPeerstore/SetAddrs-100Addrs-Caching-8                    6099           6092           -0.11%
BenchmarkBadgerDsPeerstore/GetAddrs-1Addrs-Caching-8                      11             7              -36.36%
BenchmarkBadgerDsPeerstore/GetAddrs-10Addrs-Caching-8                     11             7              -36.36%
BenchmarkBadgerDsPeerstore/GetAddrs-100Addrs-Caching-8                    11             7              -36.36%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-1Addrs-Caching-8           3497           5450           +55.85%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-10Addrs-Caching-8          23009          23771          +3.31%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-100Addrs-Caching-8         77939          77388          -0.71%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-1Addrs-Caching-8         39135          39136          +0.00%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-10Addrs-Caching-8        170120         170121         +0.00%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-100Addrs-Caching-8       1610125        1610119        -0.00%
BenchmarkBadgerDsPeerstore/AddAddrs-1Addrs-Cacheless-8                    127            124            -2.36%
BenchmarkBadgerDsPeerstore/AddAddrs-10Addrs-Cacheless-8                   677            673            -0.59%
BenchmarkBadgerDsPeerstore/AddAddrs-100Addrs-Cacheless-8                  6099           6092           -0.11%
BenchmarkBadgerDsPeerstore/SetAddrs-1Addrs-Cacheless-8                    127            124            -2.36%
BenchmarkBadgerDsPeerstore/SetAddrs-10Addrs-Cacheless-8                   677            673            -0.59%
BenchmarkBadgerDsPeerstore/SetAddrs-100Addrs-Cacheless-8                  6099           6092           -0.11%
BenchmarkBadgerDsPeerstore/GetAddrs-1Addrs-Cacheless-8                    112            77             -31.25%
BenchmarkBadgerDsPeerstore/GetAddrs-10Addrs-Cacheless-8                   477            379            -20.55%
BenchmarkBadgerDsPeerstore/GetAddrs-100Addrs-Cacheless-8                  1380           1772           +28.41%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-1Addrs-Cacheless-8         6648           6658           +0.15%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-10Addrs-Cacheless-8        44489          44378          -0.25%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-100Addrs-Cacheless-8       110894         107717         -2.86%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-1Addrs-Cacheless-8       39135          39135          +0.00%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-10Addrs-Cacheless-8      170120         170120         +0.00%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-100Addrs-Cacheless-8     1610119        1610122        +0.00%

benchmark                                                                 old bytes     new bytes     delta
BenchmarkBadgerDsPeerstore/AddAddrs-1Addrs-Caching-8                      10786         9944          -7.81%
BenchmarkBadgerDsPeerstore/AddAddrs-10Addrs-Caching-8                     34326         33667         -1.92%
BenchmarkBadgerDsPeerstore/AddAddrs-100Addrs-Caching-8                    278240        273577        -1.68%
BenchmarkBadgerDsPeerstore/SetAddrs-1Addrs-Caching-8                      10164         9943          -2.17%
BenchmarkBadgerDsPeerstore/SetAddrs-10Addrs-Caching-8                     34323         33667         -1.91%
BenchmarkBadgerDsPeerstore/SetAddrs-100Addrs-Caching-8                    278269        273561        -1.69%
BenchmarkBadgerDsPeerstore/GetAddrs-1Addrs-Caching-8                      464           272           -41.38%
BenchmarkBadgerDsPeerstore/GetAddrs-10Addrs-Caching-8                     608           416           -31.58%
BenchmarkBadgerDsPeerstore/GetAddrs-100Addrs-Caching-8                    2240          2048          -8.57%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-1Addrs-Caching-8           120265        182528        +51.77%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-10Addrs-Caching-8          752611        777324        +3.28%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-100Addrs-Caching-8         2600536       2589311       -0.43%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-1Addrs-Caching-8         2684315       2684482       +0.01%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-10Addrs-Caching-8        7321860       7321976       +0.00%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-100Addrs-Caching-8       62042404      62041576      -0.00%
BenchmarkBadgerDsPeerstore/AddAddrs-1Addrs-Cacheless-8                    10164         9943          -2.17%
BenchmarkBadgerDsPeerstore/AddAddrs-10Addrs-Cacheless-8                   34330         33668         -1.93%
BenchmarkBadgerDsPeerstore/AddAddrs-100Addrs-Cacheless-8                  278238        273548        -1.69%
BenchmarkBadgerDsPeerstore/SetAddrs-1Addrs-Cacheless-8                    10164         9943          -2.17%
BenchmarkBadgerDsPeerstore/SetAddrs-10Addrs-Cacheless-8                   34325         33668         -1.91%
BenchmarkBadgerDsPeerstore/SetAddrs-100Addrs-Cacheless-8                  278226        273566        -1.67%
BenchmarkBadgerDsPeerstore/GetAddrs-1Addrs-Cacheless-8                    6360          4679          -26.43%
BenchmarkBadgerDsPeerstore/GetAddrs-10Addrs-Cacheless-8                   24985         20777         -16.84%
BenchmarkBadgerDsPeerstore/GetAddrs-100Addrs-Cacheless-8                  75801         89835         +18.51%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-1Addrs-Cacheless-8         233188        233330        +0.06%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-10Addrs-Cacheless-8        1453068       1449961       -0.21%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-100Addrs-Cacheless-8       3677443       3582483       -2.58%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-1Addrs-Cacheless-8       2684390       2684395       +0.00%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-10Addrs-Cacheless-8      7321831       7321852       +0.00%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-100Addrs-Cacheless-8     62041620      62041938      +0.00%
```